# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TMS API",
    "version": "1.0.0",
    "description": "\n    # TMS (Training Management System) API\n    \n    Comprehensive API for managing training programs, students, companies, and security controls.\n    \n    ## Features\n    - 🏢 **Company Management**: Manage companies and their employees\n    - 👨‍🎓 **Student Management**: Handle student registration and tracking\n    - 📋 **Contract Management**: Create and manage training contracts\n    - 🔒 **Security Controls**: Safety inspections and compliance tracking\n    - 📊 **Quiz System**: Training assessments and evaluations\n    - 🚗 **Vehicle Management**: Fleet and vehicle tracking\n    - 📈 **Reporting**: Comprehensive reporting and analytics\n    \n    ## Authentication\n    Most endpoints require authentication using JWT tokens.\n    Include the token in the Authorization header:\n    `Authorization: Bearer <your-jwt-token>`\n    \n    ## Rate Limiting\n    - General API: 1000 requests per 15 minutes\n    - Authentication endpoints: 5 requests per 15 minutes\n    \n    ## Error Handling\n    All endpoints return consistent error responses with:\n    - Error message in Norwegian\n    - HTTP status codes\n    - Request ID for tracking\n    \n    ## API Versioning\n    Current version: v1\n    - Path versioning: `/api/v1/...`\n    - Header versioning: `Accept-Version: 1.0`\n    ",
    "termsOfService": "https://tms.example.com/terms",
    "contact": {
...
    },
    "license": {
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "\n        JWT token for authentication. Obtain a token by logging in via /api/auth/login.\n        \n        **Usage:**\n        `Authorization: Bearer <your-jwt-token>`\n        \n        **Token expiry:** 7 days for regular users, 30 days for admin users.\n        \n        **Refresh:** Tokens cannot be refreshed. Login again to get a new token.\n        "
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/tenants": {
      "get": {
        "summary": "Hent alle tenants for bruker",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Opprett ny tenant",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tenants/{tenantId}": {
      "get": {
        "summary": "Hent tenant detaljer",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Slett tenant",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tenants/{tenantId}/inviter": {
      "post": {
        "summary": "Inviter bruker til tenant",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tenants/{tenantId}/aksepter": {
      "post": {
        "summary": "Aksepter invitasjon til tenant",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tenants/{tenantId}/brukere/{userId}/rolle": {
      "put": {
        "summary": "Oppdater bruker rolle i tenant",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/tenants/{tenantId}/brukere/{userId}": {
      "delete": {
        "summary": "Fjern bruker fra tenant",
        "tags": [
          "Tenants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  }
}